connect with socket

+++ controller/likes_controller.js
@@ -4,32 +4,37 @@ const Comment = require('../models/comment');
 
 module.exports.toggleLike = async function(req, res) {
     try {
-        // likes/toggle/?id=abcdef&type=Post
         let likeable;
         let deleted = false;
 
-        if (req.query.type == 'Post') {
+        // Determine whether to fetch a Post or Comment based on req.query.type
+        if (req.query.type === 'Post') {
             likeable = await Post.findById(req.query.id).populate('likes');
         } else {
             likeable = await Comment.findById(req.query.id).populate('likes');
         }
 
-        // check if a like already exists
+        if (!likeable) {
+            return res.status(404).json({
+                message: "Likeable object not found"
+            });
+        }
+
+        // Check if a like already exists for the user
         let existingLike = await Like.findOne({
             likeable: req.query.id,
             onModel: req.query.type,
             user: req.user._id
         });
 
-        // if a like already exists then delete it
+        // If a like already exists, delete it; otherwise, create a new like
         if (existingLike) {
             likeable.likes.pull(existingLike._id);
-            likeable.save();
+            await likeable.save();
 
             await existingLike.deleteOne(); // Use deleteOne instead of remove
             deleted = true;
         } else {
-            // else make a new like
             let newLike = await Like.create({
                 user: req.user._id,
                 likeable: req.query.id,
@@ -37,10 +42,10 @@ module.exports.toggleLike = async function(req, res) {
             });
 
             likeable.likes.push(newLike._id);
-            likeable.save();
+            await likeable.save();
         }
 
-        return res.json(200, {
+        return res.status(200).json({
             message: "Request successful!",
             data: {
                 deleted: deleted
@@ -48,8 +53,8 @@ module.exports.toggleLike = async function(req, res) {
         });
 
     } catch (err) {
-        console.log(err);
-        return res.json(500, {
+        console.error("Error toggling like:", err);
+        return res.status(500).json({
             message: 'Internal Server Error'
         });
     }
